restructure KGE literals
