+++ dicee/abstracts.py
@@ -6,6 +6,12 @@ from typing import List, Tuple, Union
 import random
 from abc import ABC
 import lightning
+from .models.literal import LiteralEmbeddings
+from .dataset_classes import TriplePredictionDataset, LiteralDataset
+from torch.utils.data import DataLoader
+import torch.optim as optim
+import torch.nn.functional as F
+from tqdm import tqdm
 
 
 class AbstractTrainer:
@@ -604,3 +610,241 @@ class AbstractPPECallback(AbstractCallback):
         torch.save(param_ensemble, f=f"{self.path}/trainer_checkpoint_main.pt")
         if self.sample_counter > 1:
             self.sample_counter += 1
+
+class BaseInteractiveTrainKGE:
+    """
+    Abstract/base class for training knowledge graph embedding models interactively.
+    This class provides methods for re-training KGE models and also Literal Embedding model.
+    """
+
+    # @TODO: Do we really need this ?!
+    def train_triples(self, h: List[str], r: List[str], t: List[str], labels: List[float],
+                      iteration=2, optimizer=None):
+        assert len(h) == len(r) == len(t) == len(labels)
+        # (1) From List of strings to TorchLongTensor.
+        x = torch.LongTensor(self.index_triple(h, r, t)).reshape(1, 3)
+        # (2) From List of float to Torch Tensor.
+        labels = torch.FloatTensor(labels)
+        # (3) Train mode.
+        self.set_model_train_mode()
+        if optimizer is None:
+            optimizer = optim.Adam(self.model.parameters(), lr=0.1)
+        print('Iteration starts...')
+        # (4) Train.
+        for epoch in range(iteration):
+            optimizer.zero_grad()
+            outputs = self.model(x)
+            loss = self.model.loss(outputs, labels)
+            print(f"Iteration:{epoch}\t Loss:{loss.item()}\t Outputs:{outputs.detach().mean()}")
+            loss.backward()
+            optimizer.step()
+        # (5) Eval
+        self.set_model_eval_mode()
+        with torch.no_grad():
+            x = x.to(self.model.device)
+            outputs = self.model(x)
+            loss = self.model.loss(outputs, labels)
+            print(f"Eval Mode:\tLoss:{loss.item()}")
+
+    def train_k_vs_all(self, h, r, iteration=1, lr=.001):
+        """
+        Train k vs all
+        :param head_entity:
+        :param relation:
+        :param iteration:
+        :param lr:
+        :return:
+        """
+        assert len(h) == 1
+        # (1) Construct input and output
+        out = self.construct_input_and_output_k_vs_all(h, r)
+        if out is None:
+            return
+        x, labels, idx_tails = out
+        # (2) Train mode
+        self.set_model_train_mode()
+        # (3) Initialize optimizer # SGD considerably faster than ADAM.
+        optimizer = optim.Adam(self.model.parameters(), lr=lr, weight_decay=.00001)
+
+        print('\nIteration starts.')
+        # (3) Iterative training.
+        for epoch in range(iteration):
+            optimizer.zero_grad()
+            outputs = self.model(x)
+            loss = self.model.loss(outputs, labels)
+            if len(idx_tails) > 0:
+                print(
+                    f"Iteration:{epoch}\t"
+                    f"Loss:{loss.item()}\t"
+                    f"Avg. Logits for correct tails: {outputs[0, idx_tails].flatten().mean().detach()}")
+            else:
+                print(
+                    f"Iteration:{epoch}\t"
+                    f"Loss:{loss.item()}\t"
+                    f"Avg. Logits for all negatives: {outputs[0].flatten().mean().detach()}")
+
+            loss.backward()
+            optimizer.step()
+            if loss.item() < .00001:
+                print(f'loss is {loss.item():.3f}. Converged !!!')
+                break
+        # (4) Eval mode
+        self.set_model_eval_mode()
+        with torch.no_grad():
+            outputs = self.model(x)
+            loss = self.model.loss(outputs, labels)
+        print(f"Eval Mode:Loss:{loss.item():.4f}\t Outputs:{outputs[0, idx_tails].flatten().detach()}\n")
+
+    def train(self, kg, lr=.1, epoch=10, batch_size=32, neg_sample_ratio=10, num_workers=1) -> None:
+        """ Retrained a pretrain model on an input KG via negative sampling."""
+        # (1) Create Negative Sampling Setting for training
+        print('Creating Dataset...')
+        train_set = TriplePredictionDataset(kg.train_set,
+                                            num_entities=len(kg.entity_to_idx),
+                                            num_relations=len(kg.relation_to_idx),
+                                            neg_sample_ratio=neg_sample_ratio)
+        num_data_point = len(train_set)
+        print('Number of data points: ', num_data_point)
+        train_dataloader = DataLoader(train_set, batch_size=batch_size,
+                                      #  shuffle => to have the data reshuffled at every epoc
+                                      shuffle=True, num_workers=num_workers,
+                                      collate_fn=train_set.collate_fn, pin_memory=True)
+
+        # (2) Go through valid triples + corrupted triples and compute scores.
+        # Average loss per triple is stored. This will be used  to indicate whether we learned something.
+        print('First Eval..')
+        self.set_model_eval_mode()
+        first_avg_loss_per_triple = 0
+        for x, y in train_dataloader:
+            pred = self.model(x)
+            first_avg_loss_per_triple += self.model.loss(pred, y)
+        first_avg_loss_per_triple /= num_data_point
+        print(first_avg_loss_per_triple)
+        # (3) Prepare Model for Training
+        self.set_model_train_mode()
+        optimizer = optim.Adam(self.model.parameters(), lr=lr)
+        print('Training Starts...')
+        for epoch in range(epoch):  # loop over the dataset multiple times
+            epoch_loss = 0
+            for x, y in train_dataloader:
+                # zero the parameter gradients
+                optimizer.zero_grad()
+                # forward + backward + optimize
+                outputs = self.model(x)
+                loss = self.model.loss(outputs, y)
+                epoch_loss += loss.item()
+                loss.backward()
+                optimizer.step()
+            print(f'Epoch={epoch}\t Avg. Loss per epoch: {epoch_loss / num_data_point:.3f}')
+        # (5) Prepare For Saving
+        self.set_model_eval_mode()
+        print('Eval starts...')
+        # (6) Eval model on training data to check how much an Improvement
+        last_avg_loss_per_triple = 0
+        for x, y in train_dataloader:
+            pred = self.model(x)
+            last_avg_loss_per_triple += self.model.loss(pred, y)
+        last_avg_loss_per_triple /= len(train_set)
+        print(f'On average Improvement: {first_avg_loss_per_triple - last_avg_loss_per_triple:.3f}')
+
+    def train_literals(
+        self,
+        train_file_path: str = None,
+        num_epochs: int = 100,
+        lit_lr: float = 0.001,
+        eval_litreal_preds: bool = True,
+        eval_file_path: str = None,
+        lit_normalization_type: str = "z-norm",
+        batch_size: int = 1024,
+        sampling_ratio: float = None,
+        random_seed=1,
+        loader_backend: str = "pandas",
+        freeze_entity_embeddings: bool = True,
+        gate_residual: bool = True,
+        device: str = None,
+    ):
+        """
+        Trains the Literal Embeddings model using literal data.
+
+        Args:
+            train_file_path (str): Path to the training data file.
+            num_epochs (int): Number of training epochs.
+            lit_lr (float): Learning rate for the literal model.
+            eval_litreal_preds (bool): If True, evaluate the model after training.
+            eval_file_path (str): Path to evaluation data file.
+            norm_type (str): Normalization type to use ('z-norm', 'min-max', or None).
+            batch_size (int): Batch size for training.
+            sampling_ratio (float): Ratio of training triples to use.
+            loader_backend (str): Backend for loading the dataset ('pandas' or 'rdflib').
+            freeze_entity_embeddings (bool): If True, freeze the entity embeddings during training.
+            gate_residual (bool): If True, use gate residual connections in the model.
+            device (str): Device to use for training ('cuda' or 'cpu'). If None, will use available GPU or CPU.
+        """
+        # Assign torch.seed to reproduice experiments
+        torch.manual_seed(random_seed)
+        torch.cuda.manual_seed_all(random_seed)
+
+        # Set the device for training
+        try:
+            device = torch.device(device)
+        except Exception:
+            device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+                    
+        # Prepare the dataset and DataLoader
+        literal_dataset = LiteralDataset(
+            file_path=train_file_path,
+            ent_idx=self.entity_to_idx,
+            normalization_type=lit_normalization_type,
+            sampling_ratio=sampling_ratio,
+            loader_backend=loader_backend,
+        )
+
+        self.data_property_to_idx = literal_dataset.data_property_to_idx
+
+        batch_data = DataLoader(
+            dataset=literal_dataset,
+            shuffle=True,
+            batch_size=batch_size,
+        )
+
+        # Initialize Literal Embedding model
+        literal_model = LiteralEmbeddings(
+            num_of_data_properties=literal_dataset.num_data_properties,
+            embedding_dims=self.model.embedding_dim,
+            entity_embeddings=self.model.entity_embeddings,
+            freeze_entity_embeddings=freeze_entity_embeddings,
+            gate_residual=gate_residual
+        ).to(device)
+
+        optimizer = optim.Adam(literal_model.parameters(), lr=lit_lr)
+        loss_log = {"lit_loss": []}
+        literal_model.train()
+
+        print(
+            f"Training Literal Embedding model"
+            f"using pre-trained '{self.model.name}' embeddings."
+        )
+
+        # Training loop
+        for epoch in (tqdm_bar := tqdm(range(num_epochs))):
+            epoch_loss = 0
+            for batch_x, batch_y in batch_data:
+                optimizer.zero_grad()
+                lit_entities = batch_x[:, 0].long().to(device)
+                lit_properties = batch_x[:, 1].long().to(device)
+                batch_y = batch_y.to(device)
+                yhat = literal_model(lit_entities, lit_properties)
+                lit_loss = F.l1_loss(yhat, batch_y)
+                lit_loss.backward()
+                optimizer.step()
+                epoch_loss += lit_loss
+
+            avg_epoch_loss = epoch_loss / len(batch_data)
+            tqdm_bar.set_postfix_str(f"loss_lit={lit_loss:.5f}")
+            loss_log["lit_loss"].append(avg_epoch_loss.item())
+
+        self.literal_model = literal_model
+        self.literal_dataset = literal_dataset
+        torch.save(literal_model.state_dict(), self.path + "/literal_model.pt")
+        print(f"Literal Embedding model saved to {self.path}/literal_model.pt")
+
